Added backbone for team functionality
Added the necessary member variables to gamesettings and player classes in order for teams to be functional. Also modified behavior in the vessel_outpost_event class for handling gifts, in which I added a check to treat a vessel as a gift if two players are on the same team.