Added improved visual clarity on game creation screen
Patched a visual oddity that would occur when a user chose a number of teams that was valid for their current player count, then switched to player count for which that team count was no longer valid; and was thus not displayed. Now, if the user changes the number of players without reselecting the number of teams, the number of teams will be displayed as zero until the user chooses a player count for which their original choice would be valid. This change is intended to make the game's handling of team numbers during game creation more clear to the user.